# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ src/main/java/com/born2go/lazzybee/gdatabase/server/DataServiceImpl.java
@@ -6,6 +6,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
+import org.jsoup.Jsoup;
+
 import com.born2go.lazzybee.gdatabase.client.rpc.DataService;
 import com.born2go.lazzybee.gdatabase.shared.Blog;
 import com.born2go.lazzybee.gdatabase.shared.Picture;
@@ -23,6 +25,32 @@ import com.googlecode.objectify.cmd.Query;
 @SuppressWarnings("serial")
 public class DataServiceImpl extends RemoteServiceServlet implements
 		DataService {
+	
+	private boolean verifyAdmin(String userId) {
+		if(userId != null && userId.contains("_G")) {
+			User u = ofy().load().type(User.class).filter("google_id", userId).first().now();
+			if(u != null && u.isAdmin())
+				return true;
+			else
+				return false;
+		}
+		else if(userId != null && userId.contains("_F")) {
+			User u = ofy().load().type(User.class).filter("facebook_id", userId).first().now();
+			if(u != null && u.isAdmin())
+				return true;
+			else
+				return false;
+		}
+		else
+			return false;
+	}
+	
+//	private String getPlainText(String strSrc) {
+//		String resultStr = strSrc;
+//		resultStr = resultStr.replaceAll("<figcaption>.*</figcaption>", "");
+//		resultStr = resultStr.replaceAll("&nbsp;", " ");
+//		return resultStr;
+//	}
 
 	/**
 	 * Insert new vocabulary
@@ -30,14 +58,18 @@ public class DataServiceImpl extends RemoteServiceServlet implements
 	 * @return inserted vocabulary
 	 */
 	@Override
-	public Voca insertVoca(final Voca voca) {
-		voca.setQ(voca.getQ().toLowerCase());
-		voca.setCheck(false);
-		if (verifyVoca(voca.getQ())) {
-			Key<Voca> key = ofy().save().entity(voca).now();
-			Voca v = ofy().load().key(key).now();
-			return v;
-		} else
+	public Voca insertVoca(final Voca voca, String userId) {
+		if(verifyAdmin(userId)) {
+			voca.setQ(voca.getQ().toLowerCase());
+			voca.setCheck(false);
+			if (verifyVoca(voca.getQ())) {
+				Key<Voca> key = ofy().save().entity(voca).now();
+				Voca v = ofy().load().key(key).now();
+				return v;
+			} else
+				return null;
+		}
+		else
 			return null;
 	}
 
@@ -75,20 +107,24 @@ public class DataServiceImpl extends RemoteServiceServlet implements
 	 * @return updated vocabulary, null if update fail
 	 */
 	@Override
-	public Voca updateVoca(Voca voca, boolean isCheck) {
-		Voca v = ofy().load().type(Voca.class).id(voca.getGid()).now();
-		if (v != null) {
-			if (voca.getQ().equals(v.getQ())) {
-				if (isCheck)
-					voca.setCheck(true);
-				ofy().save().entity(voca);
-				return voca;
-			} else {
-				voca.setGid(null);
-				return insertVoca(voca);
+	public Voca updateVoca(Voca voca, boolean isCheck, String userId) {
+		if(verifyAdmin(userId)) {
+			Voca v = ofy().load().type(Voca.class).id(voca.getGid()).now();
+			if (v != null) {
+				if (voca.getQ().equals(v.getQ())) {
+					if (isCheck)
+						voca.setCheck(true);
+					ofy().save().entity(voca);
+					return voca;
+				} else {
+					voca.setGid(null);
+					return insertVoca(voca, userId);
+				}
 			}
+			return null;
 		}
-		return null;
+		else
+			return null;
 	}
 
 	/**
@@ -157,28 +193,42 @@ public class DataServiceImpl extends RemoteServiceServlet implements
 	 * Remove a vocabulary
 	 */
 	@Override
-	public void removeVoca(Voca voca) {
-		Voca v = ofy().load().type(Voca.class).id(voca.getGid()).now();
-		if (v != null)
-			ofy().delete().entity(v);
+	public void removeVoca(Voca voca, String userId) {
+		if(verifyAdmin(userId)) {
+			Voca v = ofy().load().type(Voca.class).id(voca.getGid()).now();
+			if (v != null)
+				ofy().delete().entity(v);
+		}
 	}
 
 	@Override
-	public void saveUser(User user) {
+	public User saveUser(User user) {
+		user.setAdmin(false);
 		if (user.getGoogle_id() != null) {
-			int i = ofy().load().type(User.class)
-					.filter("google_id", user.getGoogle_id()).count();
-			if (i == 0)
+			User old_user = ofy().load().type(User.class)
+					.filter("google_id", user.getGoogle_id()).first().now();
+			if (old_user == null) {
 				ofy().save().entity(user);
+				return user;
+			}
+			else
+				return old_user;
 		}
-		if (user.getFacebook_id() != null) {
-			int i = ofy().load().type(User.class)
-					.filter("facebook_id", user.getFacebook_id()).count();
-			if (i == 0)
+		else if (user.getFacebook_id() != null) {
+			User old_user = ofy().load().type(User.class)
+					.filter("facebook_id", user.getFacebook_id()).first().now();
+			if (old_user == null) {
 				ofy().save().entity(user);
+				return user;
+			}
+			else
+				return old_user;
 		}
+		else
+			return user;
 	}
 
+	@Override
 	public boolean verifyBlog(String blogTitle) {
 		Blog blog = ofy().load().type(Blog.class).filter("title", blogTitle)
 				.first().now();
@@ -189,25 +239,45 @@ public class DataServiceImpl extends RemoteServiceServlet implements
 	}
 
 	@Override
-	public Blog insertBlog(Blog blog) {
-		String origin_title = blog.getTitle();
-		boolean verify_blog = false;
-		for (int i = 0; i <= 9; i++) {
-			if (!verifyBlog(blog.getTitle())) {
-				Random generator = new Random();
-				int random_title = generator.nextInt(100);
-				blog.setTitle(origin_title + "_" + random_title);
-			} else {
-				verify_blog = true;
-				break;
+	public Blog insertBlog(Blog blog, String userId) {
+//		String origin_title = blog.getTitle();
+//		boolean verify_blog = false;
+//		for (int i = 0; i <= 9; i++) {
+//			if (!verifyBlog(blog.getTitle())) {
+//				Random generator = new Random();
+//				int random_title = generator.nextInt(100);
+//				blog.setTitle(origin_title + "_" + random_title);
+//			} else {
+//				verify_blog = true;
+//				break;
+//			}
+//		}
+		if(verifyAdmin(userId)) {
+			if (verifyBlog(blog.getTitle())) {
+				blog.setCreateDate(System.currentTimeMillis());
+				Key<Blog> key = ofy().save().entity(blog).now();
+				Blog b = ofy().load().key(key).now();
+				return b;
+			} else
+				return null;
 			}
+		else
+			return null;
+	}
+
+	@Override
+	public Blog updateBlog(Blog blog, String userId) {
+		if(verifyAdmin(userId)) {
+			Blog old_blog = findBlogById(blog.getId());
+			if(old_blog != null) {
+//				String preshortHtml = getPlainText(blog.getContent());
+//				blog.setShortDescription(Jsoup.parse(preshortHtml).text());
+				ofy().save().entity(blog);
+				return blog;
+			}
+			return null;
 		}
-		if (verify_blog) {
-			blog.setCreateDate(System.currentTimeMillis());
-			Key<Blog> key = ofy().save().entity(blog).now();
-			Blog b = ofy().load().key(key).now();
-			return b;
-		} else
+		else
 			return null;
 	}
 
@@ -225,26 +295,32 @@ public class DataServiceImpl extends RemoteServiceServlet implements
 	}
 
 	@Override
+	public List<Blog> getListBlog() {
+		List<Blog> blogs = ofy().load().type(Blog.class).limit(6).list();
+		List<Blog> result = new ArrayList<Blog>();
+		result.addAll(blogs);
+		return result;
+	}
+
+	@Override
 	public Picture findPicture(Long pictureId) {
-		Picture picture = ofy().load().type(Picture.class).id(pictureId).now();
-		return picture;
+		if(pictureId != null) {
+			Picture picture = ofy().load().type(Picture.class).id(pictureId).now();
+			return picture;
+		}
+		else 
+			return null;
 	}
 
 	private BlobstoreService blobStoreService = BlobstoreServiceFactory
 			.getBlobstoreService();
 
 	@Override
-	public String getUploadUrl() {
-		return blobStoreService.createUploadUrl("/photo_upload");
+	public String getUploadUrl(String userId) {
+		if(verifyAdmin(userId)) {		
+			return blobStoreService.createUploadUrl("/photo_upload");
+		}
+		else
+			return null;
 	}
-
-	@Override
-	public List<Blog> findBlogs() {
-		List<Blog> result = new ArrayList<Blog>();
-		List<Blog> list = ofy().load().type(Blog.class).limit(5).list();
-		if (list != null && !list.isEmpty())
-			result.addAll(list);
-		return result;
-	}
-
 }
